# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Svloon",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/salon-schedules": {
            "get": {
                "tags": [
                    "SalonSchedule"
                ],
                "summary": "getSalonScheduleList",
                "description": "Get all SalonSchedules",
                "operationId": "b014bbbcc62f30c0c26b63971692f2d5",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SalonSchedule"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "SalonSchedule"
                ],
                "summary": "createSalonSchedule",
                "description": "Create SalonSchedule",
                "operationId": "d27c4ebfb79e058f903830e8a07abb1c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SalonSchedule"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SalonSchedule"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/salon-schedules/{id}": {
            "get": {
                "tags": [
                    "SalonSchedule"
                ],
                "summary": "getSalonScheduleItem",
                "description": "Get SalonSchedule",
                "operationId": "59f0f47776f1ad664d69dfa99d9b056f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of SalonSchedule",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SalonSchedule"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "SalonSchedule"
                ],
                "summary": "updateSalonSchedule",
                "description": "Update SalonSchedule",
                "operationId": "f1b0813a2c431fb9470bcf3bdb5853f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of SalonSchedule",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SalonSchedule"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SalonSchedule"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "SalonSchedule"
                ],
                "summary": "deleteSalonSchedule",
                "description": "Delete SalonSchedule",
                "operationId": "f3982a37e6cfbf85ab06a245617f7f4f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of SalonSchedule",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SalonSchedule": {
                "required": [],
                "properties": {
                    "start_day": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": false,
                        "nullable": true
                    },
                    "end_dat": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": false,
                        "nullable": true
                    },
                    "start_hour": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": false,
                        "nullable": true
                    },
                    "end_hour": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": false,
                        "nullable": true
                    },
                    "pause_start": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": false,
                        "nullable": true
                    },
                    "pause_end": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": false,
                        "nullable": true
                    },
                    "is_active": {
                        "description": "",
                        "type": "boolean",
                        "readOnly": false,
                        "nullable": true
                    },
                    "created_at": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    },
                    "updated_at": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    }
}